+++ lib/domain/record/components/supports/components/rest_duration/begin_manual_rest_duration_button.dart
@@ -61,47 +61,3 @@ class BeginManualRestDurationButton extends StatelessWidget {
     );
   }
 }
-
-class EndManualRestDurationButton extends StatelessWidget {
-  final RestDuration restDuration;
-  final PillSheet activedPillSheet;
-  final PillSheetGroup pillSheetGroup;
-  final RecordPageStore store;
-
-  const EndManualRestDurationButton({
-    Key? key,
-    required this.restDuration,
-    required this.activedPillSheet,
-    required this.pillSheetGroup,
-    required this.store,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 80,
-      child: SmallAppOutlinedButton(
-        text: "休薬終了",
-        onPressed: () async {
-          analytics.logEvent(
-            name: "end_manual_rest_duration_pressed",
-          );
-
-          await store.endResting(
-            pillSheetGroup: pillSheetGroup,
-            activedPillSheet: activedPillSheet,
-            restDuration: restDuration,
-          );
-          ScaffoldMessenger.of(context).showSnackBar(
-            SnackBar(
-              duration: Duration(
-                seconds: 2,
-              ),
-              content: Text("休薬期間が終了しました"),
-            ),
-          );
-        },
-      ),
-    );
-  }
-}
