+++ lib/domain/record/components/supports/components/rest_duration/end_manual_rest_duration_button.dart
@@ -1,0 +1,50 @@
+import 'package:flutter/material.dart';
+import 'package:pilll/analytics.dart';
+import 'package:pilll/components/atoms/buttons.dart';
+import 'package:pilll/domain/record/record_page_store.dart';
+import 'package:pilll/entity/pill_sheet.dart';
+import 'package:pilll/entity/pill_sheet_group.dart';
+
+class EndManualRestDurationButton extends StatelessWidget {
+  final RestDuration restDuration;
+  final PillSheet activedPillSheet;
+  final PillSheetGroup pillSheetGroup;
+  final RecordPageStore store;
+
+  const EndManualRestDurationButton({
+    Key? key,
+    required this.restDuration,
+    required this.activedPillSheet,
+    required this.pillSheetGroup,
+    required this.store,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      width: 80,
+      child: SmallAppOutlinedButton(
+        text: "休薬終了",
+        onPressed: () async {
+          analytics.logEvent(
+            name: "end_manual_rest_duration_pressed",
+          );
+
+          await store.endResting(
+            pillSheetGroup: pillSheetGroup,
+            activedPillSheet: activedPillSheet,
+            restDuration: restDuration,
+          );
+          ScaffoldMessenger.of(context).showSnackBar(
+            SnackBar(
+              duration: Duration(
+                seconds: 2,
+              ),
+              content: Text("休薬期間が終了しました"),
+            ),
+          );
+        },
+      ),
+    );
+  }
+}
